Notepad
This is a notepad Clone which i created using JAVA Swing. In this Notepad you can get the basic functionality of the windows notepad but also you can change the Background and Font colour which is inbuilt in it. Also it has find and replace functionality to make any text file easy to use and edit. For taking hard copy of the text file , you can also print it directly using Notepad. It is a Advance version of Windows Notepad Software.